// src/api/generated/models/companySummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { CompanySummaryStatus } from './companySummaryStatus';

export interface CompanySummary {
  id?: string;
  name?: string;
  code?: string;
  logoUrl?: string;
  status?: CompanySummaryStatus;
  createdAt?: string;
  updatedAt?: string;
}