// src/api/generated/models/companySummaryStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export type CompanySummaryStatus = typeof CompanySummaryStatus[keyof typeof CompanySummaryStatus];


export const CompanySummaryStatus = {
  ACTIVE: 'ACTIVE',
  SUSPENDED: 'SUSPENDED',
  PASSIVE: 'PASSIVE',
} as const;